Removed: functionality
else statements removed that fill columns with 0s if smiles are not found

Hammett sigma summation if else statements

Comment:
These removed changes were implemented into the cleanframe functionality.

htbatch's primary functionality is to run htdesc on smiles present in .csv file.

The code is now stored in the commit history if it's necessary to re-implement these deletions into htbatch